backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib[bcrypt]
psycopg2-binary
